+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNEXT
 - Define admin addresses for deployment (#34)
+- Mark assembly blocks as memory-safe and update deal price calculation syntax. (#33)
 - Harmonize SRLC wording. (#31)
 - Rename roles. (#29)
 - Make sponsored & non-sponsored amounts always divisible by deal volume to refund tasks fairly. (#28)